+++ pkg/artifactory/datasource/repository/virtual/datasource_artifactory_virtual_repository_test.go
@@ -36,7 +36,7 @@ func TestAccDataSourceVirtualAllGenericLikeRetrievalPackageTypes(t *testing.T) {
 }
 
 func TestAccDataSourceVirtualAllGradleLikePackageTypes(t *testing.T) {
-	for _, packageType := range repository.GradleLikePackageTypes {
+	for _, packageType := range repository.PackageTypesLikeGradle {
 		t.Run(packageType, func(t *testing.T) {
 			resource.Test(mkNewVirtualTestCase(packageType, t, map[string]interface{}{
 				"description": fmt.Sprintf("%s virtual repository public description testing.", packageType),