+++ pkg/artifactory/resource/repository/federated/resource_artifactory_federated_repository_test.go
@@ -1391,7 +1391,7 @@ func makeFederatedGradleLikeRepoTestCase(repoType string, t *testing.T) (*testin
 }
 
 func TestAccFederatedAllGradleLikePackageTypes(t *testing.T) {
-	for _, packageType := range repository.GradleLikePackageTypes {
+	for _, packageType := range repository.PackageTypesLikeGradle {
 		t.Run(packageType, func(t *testing.T) {
 			resource.Test(makeFederatedGradleLikeRepoTestCase(packageType, t))
 		})